fix(plc4go): widen length/position arithmetic to close the f009 uint16 wraps, and regenerate
Completes the f009 fix at the source (the runtime guard in the ADS codec was
only the demonstrated-leg stopgap):

- Length arithmetic: GetLengthInBits/GetLengthInBytes widened uint16 -> uint64
  in spi/utils/LengthAware and the Go complex-type/enum templates, so serialized
  bodies over ~8KB can no longer wrap the length written into wire headers
  (frame smuggling). Every consumer already casts to its wire field type.
- Position arithmetic: PositionAware/PositionMover/ReadBuffer GetPos/SetPos/Reset
  widened uint16 -> uint32 across the runtime buffers and field readers, so
  position tracking no longer wraps at 64KB. The single mspec comparison that
  relied on curPos matching a uint16 wire field (S7 COTPPacket 'curPos < cotpLen')
  is fixed by widening that type parameter to uint 32.
- f017: added a 'strLen >= 1' validation to AmsString in ads-discovery.mspec so
  the vstring length expression can no longer underflow (belt-and-suspenders with
  the runtime allocation cap already in place).

Then regenerated all plc4go protocol sources and mocks from the updated
templates/mspec. Hand-updated the legacy [constants] and external-source enum
files (Ads/Bacnet/CBus/Eip/Modbus Constants, OpcuaNodeIdServicesVariable*), which
the current constants/enum generation path no longer overwrites, plus the
hand-written eip/bacnetip/cbus helpers and the position/length unit tests.

+++ plc4go/internal/eip/Reader.go
@@ -194,7 +194,7 @@ func toAnsi(tag string) ([]byte, error) {
 			}
 			newSegment = readWriteModel.NewDataSegment(readWriteModel.NewAnsiExtendedSymbolSegment(identifier, pad))
 		}
-		lengthInBytes += newSegment.GetLengthInBytes(ctx)
+		lengthInBytes += uint16(newSegment.GetLengthInBytes(ctx))
 		segments = append(segments, newSegment)
 	}
 	buffer := utils.NewWriteBufferByteBased(
