+++ plc4go/protocols/ads/discovery/readwrite/model/AmsNetId.go
@@ -235,8 +235,8 @@ func (m *_AmsNetId) GetPlx4xTypeName() string {
 	return "AmsNetId"
 }
 
-func (m *_AmsNetId) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(0)
+func (m *_AmsNetId) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(0)
 
 	// Simple field (octet1)
 	lengthInBits += 8
@@ -259,7 +259,7 @@ func (m *_AmsNetId) GetLengthInBits(ctx context.Context) uint16 {
 	return lengthInBits
 }
 
-func (m *_AmsNetId) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_AmsNetId) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
