+++ plc4go/protocols/ads/readwrite/model/AdsStampHeader.go
@@ -196,8 +196,8 @@ func (m *_AdsStampHeader) GetPlx4xTypeName() string {
 	return "AdsStampHeader"
 }
 
-func (m *_AdsStampHeader) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(0)
+func (m *_AdsStampHeader) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(0)
 
 	// Simple field (timestamp)
 	lengthInBits += 64
@@ -216,7 +216,7 @@ func (m *_AdsStampHeader) GetLengthInBits(ctx context.Context) uint16 {
 	return lengthInBits
 }
 
-func (m *_AdsStampHeader) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_AdsStampHeader) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
