+++ plc4go/protocols/ads/readwrite/model/AdsWriteResponse.go
@@ -216,8 +216,8 @@ func (m *_AdsWriteResponse) GetPlx4xTypeName() string {
 	return "AdsWriteResponse"
 }
 
-func (m *_AdsWriteResponse) GetLengthInBits(ctx context.Context) uint16 {
-	lengthInBits := uint16(m.AmsPacketContract.(*_AmsPacket).getLengthInBits(ctx))
+func (m *_AdsWriteResponse) GetLengthInBits(ctx context.Context) uint64 {
+	lengthInBits := uint64(m.AmsPacketContract.(*_AmsPacket).getLengthInBits(ctx))
 
 	// Simple field (result)
 	lengthInBits += 32
@@ -225,7 +225,7 @@ func (m *_AdsWriteResponse) GetLengthInBits(ctx context.Context) uint16 {
 	return lengthInBits
 }
 
-func (m *_AdsWriteResponse) GetLengthInBytes(ctx context.Context) uint16 {
+func (m *_AdsWriteResponse) GetLengthInBytes(ctx context.Context) uint64 {
 	return m.GetLengthInBits(ctx) / 8
 }
